refactor Carousel
