--- packages/react-cap-theme/src/components/react-carousel/components/CarouselAutoplayButton/renderCarouselAutoplayButton.tsx
@@ -1,14 +1,0 @@
-import type { JSXElement } from '@fluentui/react-utilities';
-import { assertSlots } from '@fluentui/react-utilities';
-import { renderButton } from '../../../react-button/components/Button/renderButton';
-import type {
-  CarouselAutoplayButtonSlots,
-  CarouselAutoplayButtonState,
-} from './CarouselAutoplayButton.types';
-
-export const renderCarouselAutoplayButton = (
-  state: CarouselAutoplayButtonState
-): JSXElement => {
-  assertSlots<CarouselAutoplayButtonSlots>(state);
-  return renderButton(state);
-};
