+++ packages/react-cap-theme/src/components/react-carousel/components/CarouselButton/CarouselButton.types.ts
@@ -5,10 +5,7 @@ import type {
 } from '@fluentui/react-carousel';
 import type { ComponentProps, ComponentState } from '@fluentui/react-utilities';
 import type { ButtonSlots } from '@fluentui/react-button';
-import type {
-  ButtonProps,
-  ButtonState,
-} from '../../../react-button/components/Button/Button.types';
+import type { ButtonProps, ButtonState } from '../../../react-button';
 
 export type CarouselButtonSlots = ButtonSlots &
   Pick<FluentCarouselButtonSlots, 'root'>;
