--- packages/react-cap-theme/src/components/react-carousel/components/CarouselButton/renderCarouselButton.tsx
@@ -1,14 +1,0 @@
-import type { JSXElement } from '@fluentui/react-utilities';
-import { assertSlots } from '@fluentui/react-utilities';
-import { renderButton } from '../../../react-button/components/Button/renderButton';
-import type {
-  CarouselButtonSlots,
-  CarouselButtonState,
-} from './CarouselButton.types';
-
-export const renderCarouselButton = (
-  state: CarouselButtonState
-): JSXElement => {
-  assertSlots<CarouselButtonSlots>(state);
-  return renderButton(state);
-};
