--- packages/react-cap-theme/src/components/react-carousel/components/CarouselNav/useCarouselNavContextValues.ts
@@ -1,14 +1,0 @@
-import * as React from 'react';
-import type {
-  CarouselNavContextValues,
-  CarouselNavState,
-} from './CarouselNav.types';
-
-export const useCarouselNavContextValues = (
-  state: CarouselNavState
-): CarouselNavContextValues => {
-  const { appearance } = state;
-  const carouselNav = React.useMemo(() => ({ appearance }), [appearance]);
-
-  return { carouselNav };
-};
