+++ packages/react-cap-theme/src/components/react-carousel/components/CarouselNavContainer/CarouselNavContainer.types.ts
@@ -8,7 +8,7 @@ import type {
   ComponentState,
   Slot,
 } from '@fluentui/react-utilities';
-import type { TooltipProps } from '../../../react-tooltip/components/Tooltip/Tooltip.types';
+import type { TooltipProps } from '../../../react-tooltip';
 import type { CarouselState } from '../Carousel/Carousel.types';
 
 export type CarouselNavContainerSlots = Omit<
