+++ packages/react-cap-theme/src/components/react-carousel/index.ts
@@ -1,7 +1,18 @@
 export { useCarouselStyles } from './components/Carousel/useCarouselStyles.styles';
+export type { CarouselState } from './components/Carousel/Carousel.types';
+
 export { useCarouselAutoplayButtonStyles } from './components/CarouselAutoplayButton/useCarouselAutoplayButtonStyles.styles';
+export type { CarouselAutoplayButtonState } from './components/CarouselAutoplayButton/CarouselAutoplayButton.types';
+
 export { useCarouselButtonStyles } from './components/CarouselButton/useCarouselButtonStyles.styles';
+export type { CarouselButtonState } from './components/CarouselButton/CarouselButton.types';
+
 export { useCarouselNavStyles } from './components/CarouselNav/useCarouselNavStyles.styles';
+export type { CarouselNavState } from './components/CarouselNav/CarouselNav.types';
+
 export { useCarouselNavButtonStyles } from './components/CarouselNavButton/useCarouselNavButtonStyles.styles';
+
 export { useCarouselNavContainerStyles } from './components/CarouselNavContainer/useCarouselNavContainerStyles.styles';
+export type { CarouselNavContainerState } from './components/CarouselNavContainer/CarouselNavContainer.types';
+
 export { useCarouselNavImageButtonStyles } from './components/CarouselNavImageButton/useCarouselNavImageButtonStyles.styles';
