+++ packages/react-cap-theme/src/components/react-label/index.ts
@@ -1,1 +1,2 @@
 export { useLabelStyles } from './components/Label/useLabelStyles.styles';
+export type { LabelState } from './components/Label/Label.types';
