+++ packages/react-cap-theme/src/components/react-tooltip/index.ts
@@ -1,2 +1,5 @@
 export { useTooltipStyles_unstable as useTooltipStyles } from './components/Tooltip/useTooltipStyles.styles';
-export type { TooltipState } from './components/Tooltip/Tooltip.types';
+export type {
+  TooltipProps,
+  TooltipState,
+} from './components/Tooltip/Tooltip.types';
